+++ sparkplug/src/drivers/lp55231driver.h
@@ -16,6 +16,7 @@ class LP55231Driver : public Lp55231, public WireDevice, public OutputDevice
     LP55231Driver(uint8_t address_ = 0x00, uint16_t channelFrom_ = 0, uint16_t channelCount_ = 0);
     void setup();
     void output();
+    const char *getName();
 
   private:
     static const uint8_t REG_D1_PWM = 0x16;