+++ sparkplug/src/types/wiredevice.h
@@ -8,4 +8,5 @@ struct WireDevice : Device
     uint8_t address;
     bool connected;
     uint8_t errorState;
+    virtual const char *getName() = 0;
 };